requirements.txt
bleach==3.2.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.4
comtypes==1.1.7
cryptography==3.1.1
docutils==0.16
idna==2.10
keyring==21.4.0
packaging==20.4
pdfminer.six==20201018
pkginfo==1.5.0.1
pycparser==2.20
Pygments==2.7.1
pyparsing==2.4.7
PyPDF2==1.26.0
pypiwin32==223
pyttsx3==2.90
pywin32==228
pywin32-ctypes==0.2.0
readme-renderer==27.0
requests==2.24.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
six==1.15.0
sortedcontainers==2.2.2
termcolor==1.1.0
tqdm==4.50.2
twine==3.2.0
urllib3==1.25.10
webencodings==0.5.1
